Hardcode the Specification class description into the rdoc class template
We only use this template to generate the specification reference. Unfortunately rdoc doesn't see the description in the source
for Specification, because it is above the class name line. This will need to be fixed upstream in the rubygems code. For now,
just extract the current description from the rubygems source to get unstuck.