Update instructions for setting up RubyGems.org account
I tried using the 'curl' command but got errors about certificates.  After searching for many hours, I found out that Anaconda had installed itself in my path and I was using a different curl version.  Also, OpenSSL that is installed with Max OS X is different than the one installed with brew.  So, I finally figured out that if I just use the URL in my browser, everything works and I can avoid the hassles of curl and security certificates.  I don't know why this is happening and I don't know what certificate is used by the RubyGems.org server.  But it seems that the browser is perfectly happy with the certificate, however curl complains about the certificate.