Add RubyGems Basics page
This covers basics of the `gem` command and how to require code.  A TOC
may be warranted, but it should come from a plugin (assuming there is
one).

Moved part of the "What is a gem" page into this the "Basics" page and
made it second in the list.

Moved part of the "What is a gem" page to the index to give visitors a
better idea about what this whole guides thing is about with some
appropriate styling.