Point docs.rubygems.org at guides.rubygems.org
The disk hosting docs.rubygems.org has died.  This is OK because on
Monday (2013-09-08) RubyGems 2.1 will be released and the guides site
will have the same information as the old docs site (but not the same
exact content).

Yes, this will break URLs.  We'll probably need to do something about
that.