Clearer wording on optimistic and pessimistic versioning
The current wording provides an ambiguous (and incorrect, in one case) interpretation of how optimistic versioning works.

Specifically:

> This is an "optimistic" version constraint. It's saying that all changes from 2.x on will work with my software, but for version 3.0.0 this will not be true.

I realize that this is in reference to why one might want a pessimistic constraint, but because it is listed under the optimistic constraint section, it implies that an optimistic constraint will not allow for version 3.0.0 to work, which is exactly the opposite of how optimistic versioning works. In my case, this caused me to think I had misunderstood the Gemspec specification, rather than found what turns out to (I think) be a bug in Bundler: https://github.com/bundler/bundler/issues/4945

I've rewritten this section in a way that I think is much clearer. Hopefully it's useful to others reading through it.